requirements.txt
pytest==8.3.4
playwright
requests
qase-pytest==4.2.0
python-dotenv
testit-adapter-pytest
testit-api-client==5.0.0
qaseio==3.2.1
junitxml==0.7